Supporting wiki attachments (closes #5)

+++ gitlab_direct/Connection.py
@@ -29,14 +29,32 @@ class Connection(object):
         database_proxy.initialize(db)
         self.uploads_path = uploads_path
 
-    def clear_database(self, project_id):
-        for note in Notes.select().where((Notes.project == project_id) & ~(Notes.attachment >> None)):
-            directory = os.path.join(self.uploads_path, 'note/attachment/%s' % note.id)
-            shutil.rmtree(directory)
-        Events.delete().where( (Events.project == project_id) & (Events.target_type << ['Issue', 'Note'] ) ).execute()
-        Notes.delete().where( (Notes.project == project_id) & (Notes.noteable_type == 'Issue') ).execute()
-        Issues.delete().where( Issues.project == project_id ).execute()
+    def clear_issues(self, project_id):
+        # Delete issues and everything that goes with them...
+        for issue in Issues.select().where(Issues.project == project_id):
+            for note in Notes.select().where( (Notes.project == project_id) & (Notes.noteable_type == 'Issue') & (Notes.noteable == issue.id)):
+                if note.attachment != None:
+                    directory = os.path.join(self.uploads_path, 'note/attachment/%s' % note.id)
+                    try:
+                        shutil.rmtree(directory)
+                    except:
+                        pass
+                Events.delete().where( (Events.project == project_id) & (Events.target_type == 'Note' ) & (Events.target == note.id) ).execute()
+                note.delete_instance()    
+            Events.delete().where( (Events.project == project_id) & (Events.target_type == 'Issue' ) & (Events.target == issue.id) ).execute()
+            issue.delete_instance()
+            
         Milestones.delete().where( Milestones.project == project_id ).execute()
+
+    def clear_wiki_attachments(self, project_id):
+        for note in Notes.select().where( (Notes.project == project_id) & (Notes.noteable_type >> None) & (Notes.note % 'Wiki attachment %')):
+            directory = os.path.join(self.uploads_path, 'note/attachment/%s' % note.id)
+            try:
+                shutil.rmtree(directory)
+            except:
+                    pass
+            Events.delete().where( (Events.project == project_id) & (Events.target_type == 'Note' ) & (Events.target == note.id) ).execute()
+            note.delete_instance()    
 
     def milestone_by_name(self, project_id, milestone_name):
         for milestone in Milestones.select().where((Milestones.title == milestone_name) & (Milestones.project == project_id)):
@@ -125,7 +143,38 @@ class Connection(object):
             updated_at = note.created_at
         )
         event.save()
+        
+    def create_wiki_attachment(self, project_id, user, last_modified, path, binary):
+        note = Notes.create(
+            project = project_id,
+            note = 'Wiki attachment %s' % path,
+            user = user,
+            created_at = last_modified,
+            updated_at = last_modified,
+            attachment = path
+        )
+        note.save()
+        full_path = os.path.join(self.uploads_path, 'note/attachment/%s' % note.id, path)
+        directory = os.path.dirname(full_path)
+        if not os.path.exists(directory):
+            os.makedirs(directory)
+        file = open(full_path,"wb")
+        file.write(binary)
+        file.close()
+        
+        event = Events.create(
+            action = 1,
+            author = note.author,
+            created_at = note.created_at,
+            project = project_id,
+            target = note.id,
+            target_type = 'Note',
+            updated_at = note.created_at
+        )
+        event.save()
 
+        return '/files/note/%s/%s' % (note.id, path)
+        
 '''
 This file is part of <https://gitlab.dyomedea.com/vdv/trac-to-gitlab>.
